code optimization bug fix doc coverage
