+++ src/main/java/com/strandls/document/pojo/DocumentCoverageData.java
@@ -11,7 +11,9 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
  */
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class DocumentCoverageData {
-
+	private Long id;
+	private Long documentId;
+	private Long landscapeIds;
 	private String placename;
 	private String topology;
 	private Long geoEntityId;
@@ -28,11 +30,38 @@ public class DocumentCoverageData {
 	 * @param topology
 	 * @param geoEntityId
 	 */
-	public DocumentCoverageData(String placename, String topology, Long geoEntityId) {
+	public DocumentCoverageData(Long id, Long documentId,Long landscapeIds,String placename, String topology, Long geoEntityId) {
 		super();
+		this.id = id;
+		this.documentId = documentId;
+		this.landscapeIds = landscapeIds;
 		this.placename = placename;
 		this.topology = topology;
 		this.geoEntityId = geoEntityId;
+	}
+
+	public Long getId() {
+		return id;
+	}
+
+	public void setId(Long id) {
+		this.id = id;
+	}
+
+	public Long getDocumentId() {
+		return documentId;
+	}
+
+	public void setDocumentId(Long documentId) {
+		this.documentId = documentId;
+	}
+
+	public Long getLandscapeIds() {
+		return landscapeIds;
+	}
+
+	public void setLandscapeIds(Long landscapeIds) {
+		this.landscapeIds = landscapeIds;
 	}
 
 	public String getPlacename() {
